% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\docType{data}
\name{coord}
\alias{coord}
\title{coord}
\format{An object of class \code{"matrix"};
\describe{
\item{coord}{1597 cells x 2 axes (x and y)}
}}
\usage{
data(coord)
}
\description{
Spatial cell coordinates
}
\examples{
library(ssenet)
data(coord)
}
\references{
Zhu \emph{et al.} Identification of spatially associated subpopulations by combining scRNAseq and sequential fluorescence in situ hybridization data. \emph{Nat Biotechnol.} 2018 Oct 29.
(\href{https://www.nature.com/articles/nbt.4260}{Zhu \emph{et al} 2018})
}
\keyword{scRNAseq,}
\keyword{seqFISH,}
