% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.enet.R
\name{predict.enet}
\alias{predict.enet}
\title{iterxM-fold cross-validation function for Elastic net}
\usage{
\method{predict}{enet}(
  object,
  validation = c("Mfold", "loo"),
  M = 5,
  iter = 10,
  ncores = 4,
  progressBar = TRUE
)
}
\arguments{
\item{object}{an object of a specific machine learning method (ML) class}

\item{validation}{a string type of validation; either "Mfold" or "loo"}

\item{M}{an integer specifying the number of folds}

\item{iter}{an integer specifying the number of times to repeat the cross-validation}

\item{ncores}{an object of a specific machine learning method (ML) class}

\item{progressBar}{a boolean that specifies whether a progress bar should be displayed or not}
}
\description{
Estimate test error of elastic net panel
}
