% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{cross-validation function for all methods}
\usage{
predict(object, ...)
}
\arguments{
\item{x}{an object of a specific machine learning method (ML) class}

\item{validation}{a string type of validation; either "Mfold" or "loo"}

\item{M}{an integer specifying the number of folds}

\item{iter}{an integer specifying the number of times to repeat the cross-validation}

\item{threads}{an object of a specific machine learning method (ML) class}

\item{progressBar}{a boolean that specifies whether a progress bar should be displayed or not}
}
\description{
Estimate test error of elastic net panel
}
