% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrnaseq.R
\docType{data}
\name{scrnaseq}
\alias{scrnaseq}
\title{scrnaseq}
\format{An object of class \code{"data.frame"};
\describe{
\item{scrnaseq}{113 genes x 1723 cells}
}}
\usage{
data(scrnaseq)
}
\description{
normalized scRNAseq data
}
\examples{
library(ssenet)
data(scrnaseq)
}
\references{
Zhu \emph{et al.} Identification of spatially associated subpopulations by combining scRNAseq and sequential fluorescence in situ hybridization data. \emph{Nat Biotechnol.} 2018 Oct 29.
(\href{https://www.nature.com/articles/nbt.4260}{Zhu \emph{et al} 2018})
}
\keyword{scRNAseq,}
\keyword{seqFISH,}
