% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrnaseqLabels.R
\docType{data}
\name{scrnaseqLabels}
\alias{scrnaseqLabels}
\title{scrnaseqLabels}
\format{An object of class \code{"data.frame"};
\describe{
\item{scrnaseqLabels}{1723 cells x 3 variables}
}}
\usage{
data(scrnaseqLabels)
}
\description{
cell type labels for scRNAseq
}
\examples{
library(ssenet)
data(scrnaseqLabels)
}
\references{
Zhu \emph{et al.} Identification of spatially associated subpopulations by combining scRNAseq and sequential fluorescence in situ hybridization data. \emph{Nat Biotechnol.} 2018 Oct 29.
(\href{https://www.nature.com/articles/nbt.4260}{Zhu \emph{et al} 2018})
}
\keyword{scRNAseq,}
\keyword{seqFISH,}
