% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssenet.R
\name{ssenet}
\alias{ssenet}
\title{Semi-supervised Elastic net classification panel}
\usage{
ssenet(
  xtrain,
  ytrain,
  alpha,
  lambda = NULL,
  family,
  xtest = NULL,
  ytest = NULL,
  filter = "p.value",
  topranked = 50,
  keepVar = NULL,
  useObsWeights = FALSE,
  max.iter = 100,
  perc.full = 1,
  thr.conf = 0.5
)
}
\arguments{
\item{xtrain}{nxp matrix - training dataset}

\item{ytrain}{factor - response variable}

\item{alpha}{= 1 (lasso), alpha = 0 (ridge), 0 < alpha < 1 (elastic net penalty)}

\item{lambda}{= strength of elastic net penalty}

\item{family}{"binomial" or "multinomial"}

\item{xtest}{nxp matrx  - test dataset}

\item{ytest}{factor - response variable}

\item{filter}{= "none" or "p.value"}

\item{topranked}{= 50 (top number of features to select and build a classifier)}

\item{keepVar}{- names of specific variable to keep in model}

\item{weights}{- observational weights; default to 1}
}
\description{
build an elastic net classification panel
}
